+++ .github/workflows/automatic_deployment_pipeline.yaml
@@ -1,0 +1,25 @@
+name: Deploy a producción
+
+# on:
+#   push:
+#     branches:
+#       - production
+
+jobs:
+  deploy:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout del repo
+        uses: actions/checkout@v4
+
+      - name: Configurar Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+
+      - name: Instalar dependencias
+        run: pip install -r requirements.txt
+
+      - name: Desplegar al host
+        run: |
+          # Dejo este template, después hay que terminar las acciones de despliegue
