+++ .github/workflows/ci_pipeline.yml
@@ -1,10 +1,10 @@
 name: CI Pipeline
 
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
+# on:
+#   push:
+#     branches: [ main ]
+#   pull_request:
+#     branches: [ main ]
 
 jobs:
   build-and-test:
