+++ README.md
@@ -4,3 +4,11 @@
 
 ## Grupo 8 : Hugo Cabaña - Bruno Francisco Barra Atarama
 
+## Algunas decisiones tomadas:
+
+- Separamos la API con sus endpoint en `api.py` de `main.py` y `state_pattern.py`, que contiene rutinas provistas en la consigna y las clases con sus métodos para llevar a cabo el procesamiento de pagos con el patrón de diseño por estados.
+
+
+## Nomenclatura:
+
+`feat/`: en ramas, indica que es una rama de desarrollo de una *feature* que responda a un *issue*.
