+++ main.py
@@ -1,77 +1,84 @@
-import json 
- 
-from fastapi import FastAPI 
- 
-STATUS = "status" 
-AMOUNT = "amount" 
-PAYMENT_METHOD = "payment_method" 
- 
-STATUS_REGISTRADO = "REGISTRADO" 
-STATUS_PAGADO = "PAGADO" 
-STATUS_FALLIDO = "FALLIDO" 
- 
-DATA_PATH = "data.json" 
- 
+from fastapi import FastAPI, Query, HTTPException
+from src.base import load_all_payments, save_payment, load_payment, save_payment_data
+from src.base import STATUS, AMOUNT, PAYMENT_METHOD
+from src.base import STATUS_REGISTRADO, STATUS_FALLIDO, STATUS_PAGADO
+
+from src.payment_context import PaymentContext
+# from main import DATA_PATH
 app = FastAPI() 
- 
- 
-def load_all_payments(): 
-    with open(DATA_PATH, "r") as f: 
-        data = json.load(f) 
-    return data 
- 
- 
-def save_all_payments(data): 
-    with open(DATA_PATH, "w") as f: 
-        json.dump(data, f, indent=4) 
- 
- 
-def load_payment(payment_id): 
-    data = load_all_payments()[payment_id] 
-    return data 
- 
- 
-def save_payment_data(payment_id, data): 
-    all_data = load_all_payments() 
-    all_data[str(payment_id)] = data 
-    save_all_payments(all_data) 
- 
- 
-def save_payment(payment_id, amount, payment_method, status): 
-    data = { 
-        AMOUNT: amount, 
-        PAYMENT_METHOD: payment_method, 
 
-        STATUS: status, 
-    } 
-    save_payment_data(payment_id, data) 
+@app.get('/')
+async def root():
+    return {'message': 'Endpoint de prueba.'}
 
-'''
-# Ejemplo de uso: 
-# Actualizando el status de un pago: 
-data = load_payment(payment_id) 
-data[STATUS] = STATUS_PAGADO 
-save_payment_data(payment_id, data) 
-'''
-  
-# Endpoints a implementar: 
-# * GET en el path /payments que retorne todos los pagos. 
-# * POST en el path /payments/{payment_id} que registre un nuevo pago. 
-# * POST en el path /payments/{payment_id}/update que cambie los parametros de una pago (amount, payment_method) 
-# * POST en el path /payments/{payment_id}/pay que intente. 
-# * POST en el path /payments/{payment_id}/revert que revertir el pago. 
- 
-'''
-# Ejemplos: 
- 
-@app.get("/path/{arg_1}") 
-async def endpoint_a(arg_1: str, arg_2: float): 
-    # Este es un endpoint GET que recibe un argumento (arg_1) por path y otro por query (arg_2). 
-    return {} 
- 
-@app.post("/path/{arg_1}/some_action") 
-async def endpoint_b(arg_1: str, arg_2: float, arg_3: str): 
-    # Este es un endpoint POST que recibe un argumento (arg_1) por path y otros dos por query (arg_2 y 
-arg_3). 
-    return {} 
-'''
+@app.get('/payments')
+async def get_all_payments():
+    all_payments = load_all_payments()
+    return {'all_payments' : all_payments}
+
+@app.post('/payments/{payment_id}')
+async def register_payment(
+    payment_id: str,                  
+    amount: float = Query(...),       
+    payment_method: str = Query(...)
+):
+    
+    save_payment(payment_id=payment_id, amount=amount, payment_method=payment_method, status=STATUS_REGISTRADO)
+    data = load_payment(payment_id=payment_id)
+    return {'payment_id' : payment_id, 'data': data}
+
+@app.post('/payments/{payment_id}/update')
+async def update_payment(
+    payment_id: str,
+    amount: float = Query(...),
+    payment_method: str = Query(...),
+):
+    try:
+        data = load_payment(payment_id=payment_id)
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f'Error al cargar el pago para actualizarlo {e}.')
+    data[AMOUNT] = amount
+    data[PAYMENT_METHOD] = payment_method
+    save_payment_data(payment_id=payment_id, data=data)
+    return {'payment_id' : payment_id, 'data': data}
+
+@app.post('/payments/{payment_id}/')
+async def pay(payment_id: str):
+    try:
+        data = load_payment(payment_id=payment_id)
+        ctx = PaymentContext(
+                payment_id=payment_id,
+                amount=data["amount"],
+                payment_method=data["payment_method"],
+                status=data["status"]
+            )
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f'Error al cargar el pago {e}.')
+
+    try:
+        ctx.pay()
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f'Error al procesar el pago {e}')
+
+@app.post('/payments/{payment_id}/revert')
+async def revert(payment_id: str):
+    try:
+        data = load_payment(payment_id=payment_id)
+        ctx = PaymentContext(
+                payment_id=payment_id,
+                amount=data["amount"],
+                payment_method=data["payment_method"],
+                status=data["status"]
+            )
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f'Error al cargar el pago {e}.')
+
+    try:
+        ctx.revert()
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f'Error al revertir el pago {e}')
+
+
+
+
+
